MDEV-17958 Keep big endian variant of innodb_checksum_algorithm=crc32 only in big-endian machine
Problem:
========
In MySQL 5.7, it was noticed that files are not portable between
big-endian and little-endian systems (such as SPARC and x86),
because the original implementation of innodb_checksum_algorithm=crc32
was not byte order agnostic. Basically, it weakened the
innodb_checksum_algorithm for crc32 algorithm for mariadb-10.2

Fix:
====
Keep the bug-compatible variant for crc32 algorithm only present
on big-endian systems (#ifdef WORDS_BIGENDIAN).